Move all the phrase tests into their own file
